histogramresolution is int not double.
Former-commit-id: aa2e50bec04ce8408392926948cbc00dcc490f98
Former-commit-id: 662d93596c163c68f0b0379107515bbdf553d626

+++ NINA/Utility/AtikSDK/AtikCameraDll.cs
@@ -95,7 +95,7 @@ namespace NINA.Utility.AtikSDK {
             return ArtemisCameraState(camera);
         }
 
-        public static async Task<ImageArray> DownloadExposure(IntPtr camera, bool isBayered, double histogramResolution) {
+        public static async Task<ImageArray> DownloadExposure(IntPtr camera, bool isBayered, int histogramResolution) {
             CheckError(ArtemisGetImageData(camera, out var x, out var y, out var w, out var h, out var binX, out var binY), MethodBase.GetCurrentMethod(), camera);
 
             var ptr = ArtemisImageBuffer(camera);
