+++ NINA/Utility/XISF.cs
@@ -14,7 +14,7 @@ namespace NINA.Utility {
             this.Header = header;
         }
 
-        public static async Task<ImageArray> LoadImageArrayFromFile(Uri filePath, bool isBayered, double histogramResolution) {
+        public static async Task<ImageArray> LoadImageArrayFromFile(Uri filePath, bool isBayered, int histogramResolution) {
             using (FileStream fs = new FileStream(filePath.AbsolutePath, FileMode.Open)) {
                 byte[] arr = new byte[16];
                 fs.Read(arr, 0, 16);
